@@ -1,20 +1,16 @@
 package foundation.omni.proxy.analysis;
 
 import foundation.omni.CurrencyID;
-import foundation.omni.OmniDivisibleValue;
 import foundation.omni.json.pojo.OmniPropertyInfo;
 import foundation.omni.netapi.omnicore.RxOmniClient;
 import foundation.omni.rpc.SmartPropertyListInfo;
 import io.micronaut.context.annotation.Requires;
 import io.reactivex.rxjava3.core.Flowable;
-import io.reactivex.rxjava3.core.Observable;
+import io.reactivex.rxjava3.core.Maybe;
 import io.reactivex.rxjava3.core.Single;
 import io.reactivex.rxjava3.disposables.Disposable;
-import io.reactivex.rxjava3.internal.operators.observable.ObservableInterval;
-import org.bitcoinj.core.Sha256Hash;
 import org.bitcoinj.params.MainNetParams;
 import org.consensusj.bitcoin.json.pojo.ChainTip;
-import org.consensusj.bitcoin.json.pojo.TxOutSetInfo;
 import org.consensusj.bitcoin.rx.jsonrpc.service.TxOutSetService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -22,56 +18,59 @@ import org.slf4j.LoggerFactory;
 import jakarta.inject.Singleton;
 
 import java.io.Closeable;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
+
+import static foundation.omni.CurrencyID.OMNI;
+import static foundation.omni.CurrencyID.TOMNI;
+import static foundation.omni.CurrencyID.USDT;
 
 /**
- * Maintain a cache of {@link OmniPropertyInfo} for serving requests
+ * Maintain a cache of {@link OmniPropertyInfo} for serving requests. The cache is loaded
+ * using data received from {@link RxOmniClient}.
  */
 @Singleton
 @Requires(property="omniproxyd.enabled", value = "true")
 public class OmniPropertyListService implements Closeable {
     private static final Logger log = LoggerFactory.getLogger(OmniPropertyListService.class);
+
+    private final RxOmniClient rxJsonClient;
+    private final OmniPropertyListCache cache;
     private final List<CurrencyID> activeProperties;
-    RxOmniClient rxOmniClient;
+    private final TxOutSetService txOutSetService;
+    private final Flowable<Long> timerInterval;
+
     private Disposable chainTipSubscription;
     private Disposable intervalSubscription;
     private Disposable outSetSubscription;
 
-    private final ConcurrentHashMap<CurrencyID, OmniPropertyInfo> cachedPropertyInfo = new ConcurrentHashMap<>(2000);
-
-    private final Observable<Long> loadPollingInterval;
-
-    private final TxOutSetService txOutSetService;
-
-    OmniPropertyListService(RxOmniClient rxBitcoinClient) {
-        rxOmniClient = rxBitcoinClient;
-        txOutSetService = new TxOutSetService(rxBitcoinClient);
-        loadPollingInterval = ObservableInterval.interval(3,1, TimeUnit.SECONDS);
-        if (rxBitcoinClient.getNetParams().getId().equals(MainNetParams.ID_MAINNET)) {
-            activeProperties = List.of(CurrencyID.OMNI, CurrencyID.USDT);
-        } else {
-            activeProperties = List.of(CurrencyID.OMNI, CurrencyID.TOMNI);
-        }
-        cachedPropertyInfo.put(CurrencyID.BTC, OmniPropertyInfo.mockBitcoinPropertyInfo(rxBitcoinClient.getNetParams()));
+    OmniPropertyListService(RxOmniClient rxOmniClient) {
+        rxJsonClient = rxOmniClient;
+        cache = new OmniPropertyListCache(rxJsonClient.getNetParams());
+        activeProperties = rxJsonClient.getNetParams().equals(MainNetParams.get()) ? List.of(OMNI, TOMNI, USDT) : List.of(OMNI, TOMNI);
+        txOutSetService = new TxOutSetService(rxOmniClient);
+        timerInterval = Flowable.interval(3,1, TimeUnit.SECONDS);
     }
 
+    /**
+     * Subscribe to publishers
+     */
     public synchronized void start() {
+        // Subscribe to a new block (ChainTip) stream
         if (chainTipSubscription == null) {
             log.info("starting");
-            chainTipSubscription = Flowable.fromPublisher(rxOmniClient.chainTipPublisher()).subscribe(this::onNewBlock, this::onError);
+            chainTipSubscription = Flowable.fromPublisher(rxJsonClient.chainTipPublisher()).subscribe(this::onNewBlock, this::onError);
         }
-        if (intervalSubscription == null) {
-            intervalSubscription = loadPollingInterval.subscribe(i -> loadProperties(), this::onError);
-        }
+        // Subscribe to a TxOutSetInfo stream (happens once per block, but delayed because the calculation takes some time)
         if (outSetSubscription == null) {
             outSetSubscription = txOutSetService.getTxOutSetPublisher()
-                    .subscribe(this::updateBitcoin,
+                    .subscribe(cache::cachePutBitcoin,
                             t -> log.error("TxOutSetService", t),
                             () -> log.error("TxOutSetService completed"));
+        }
+        // Timer ticks used for resolving placeholders
+        if (intervalSubscription == null) {
+            intervalSubscription = timerInterval.subscribe(i -> loadPlaceholders(), this::onError);
         }
     }
 
@@ -82,103 +81,77 @@ public class OmniPropertyListService implements Closeable {
         outSetSubscription.dispose();
     }
 
+    /**
+     * Return the entire cache
+     * @return entire contents of the cache
+     */
     public Single<List<OmniPropertyInfo>> getProperties() {
-        // Return everything in the cache
-        List<OmniPropertyInfo> resultList = cachedPropertyInfo.values().stream().collect(Collectors.toUnmodifiableList());
-        return Single.just(resultList);
-    }
-
-    public Single<OmniPropertyInfo> getProperty(CurrencyID id) {
-        // Return a single entry from the cache
-        return Single.just(cachedPropertyInfo.get(id));
+        return Single.just(cache.getAll().toList());
     }
 
     /**
-     * On a new block, update the cache and initiate fetch requests for the Currency IDs on
-     * the "eager" list.
+     * Return a single entry from the cache
+     * @param id property to request
+     * @return a single entry from the cache
+     */
+    public Single<OmniPropertyInfo> getProperty(CurrencyID id) {
+        // Return a single entry from the cache
+        return Single.just(cache.get(id));
+    }
+
+    /**
+     * On a new block, update the cache by:
+     * <ol>
+     *      <li>initiate get-property requests for the Currency IDs on the "active" list</li>
+     *      <li>Poll the entire smart property list to look for additions</li>
+     * </ol>
+     * TODO: Find changed properties by looking at transactions in the block
      *
      * @param tip the new ChainTip (currently unused)
      */
     private void onNewBlock(ChainTip tip) {
         log.info("New Block -- updating CurrencyIDs on the eager list and fetching any new properties created");
-        updateActiveProperties();
-        // TODO: Find changed properties in the block
-        handleNewProperties();
-        //loadProperties();
+        activeProperties.forEach(this::getPropertyAsync);
+        omniListPropertiesMaybe().subscribe(list -> list.forEach(cache::cachePutIfNew), this::onError);
     }
 
-    private void updatePropertyAsync(CurrencyID id) {
-        log.debug("Fetching {}", id);
-        var disposable = rxOmniClient.pollOnce(() -> rxOmniClient.omniGetProperty(id))
-                .subscribe(info -> update(id, info), this::onError);
-    }
-
-    private void update(CurrencyID id, OmniPropertyInfo info) {
-        log.debug("Updating {}/{} TxID {} Amount {}", info.getPropertyid(), info.getName(), info.getCreationtxid(), info.getTotaltokens());
-        cachedPropertyInfo.put(id, info);
-    }
-
-    private void updateBitcoin(TxOutSetInfo outSetInfo) {
-        log.info("Block height is {}, Bitcoin Supply is now: {}", outSetInfo.getHeight(), outSetInfo.getTotalAmount().toFriendlyString());
-        update(CurrencyID.BTC, OmniPropertyInfo.bitcoinPropertyInfo(rxOmniClient.getNetParams(), outSetInfo.getTotalAmount()));
-    }
-
-    private void updateActiveProperties() {
-        activeProperties.forEach(this::updatePropertyAsync);
-    }
-
-    private void handleNewProperties() {
-        rxOmniClient.pollOnce(() -> rxOmniClient.omniListProperties())
-                .subscribe(this::onListProperties, this::onError);
-    }
-
-    private void onListProperties(List<SmartPropertyListInfo> list) {
-        final int perBlockNewFetch = 5;
-        List<CurrencyID> newProperties = new ArrayList<>();
-        list.forEach(splInfo -> {
-            if (!cachedPropertyInfo.containsKey(splInfo.getPropertyid())) {
-                newProperties.add(splInfo.getPropertyid());
-                cachedPropertyInfo.put(splInfo.getPropertyid(), new OmniPropertyInfo(rxOmniClient.getNetParams(), splInfo));
-            }
-        });
-        // Fetch up to perBlockNewFetch new properties per block
-        int subsize = Math.min(perBlockNewFetch, newProperties.size());
-        List<CurrencyID> newUpdateList = newProperties.subList(0, subsize);
-        newUpdateList.forEach(this::updatePropertyAsync);
-    }
 
     private void onError(Throwable t){
-        log.error("meh", t);
+        log.error("Fatal stream error", t);
     }
 
-    private void loadProperties() {
-        final int maxPerTickLoadFetch = 10;
-        List<CurrencyID> unloadedProperties = new ArrayList<>();
-        cachedPropertyInfo.forEach((id, info) -> {
-            if (!isLoaded(info)) {
-                unloadedProperties.add(id);
-            }
-        });
-        unloadedProperties.sort(CurrencyID::compareTo);
-        int subsize = Math.min(maxPerTickLoadFetch, unloadedProperties.size());
-        List<CurrencyID> loadList = unloadedProperties.subList(0, subsize);
-        if (unloadedProperties.size() > 0) {
-            log.info("Found {} unloaded properties, loading {} of them asynchronously", unloadedProperties.size(), loadList.size());
+    /**
+     * Load a batch of placeholder properties
+     */
+    private void loadPlaceholders() {
+        final int maxFetchPerTimerTick = 10;
+        // Find all properties needing loading
+        List<CurrencyID> placeholders = cache.getPlaceholderIds();
+        // Create a prioritized (ascending numeric order by ID) list to load on this tick
+        List<CurrencyID> loadList = placeholders.stream()
+                .sorted()
+                .limit(maxFetchPerTimerTick)
+                .toList();
+        if (loadList.size() > 0) {
+            log.info("Found {} placeholder properties, loading {}, starting with {}", placeholders.size(), loadList.size(), loadList.get(0));
         }
-        loadList.forEach(this::updatePropertyAsync);
+        loadList.forEach(this::getPropertyAsync);
     }
 
-    private static boolean isLoaded(OmniPropertyInfo info) {
-        // All tokens besides OMNI and TOMNI must have a creationTxId
-        if (info.getPropertyid().equals(CurrencyID.BTC)) {
-            return true;  // Bitcoin isn't loaded as an Omni property, so it's always "loaded"
-        } else if (info.getPropertyid().equals(CurrencyID.OMNI)) {
-            return !info.getTotaltokens().equals(OmniDivisibleValue.ZERO);
-        } else if (info.getPropertyid().equals(CurrencyID.TOMNI)) {
-            // We don't have a good test for TOMNI that works on a syncing mainnet, totalTokens stays zero until the exodus block it seems??
-            return !info.getTotaltokens().equals(OmniDivisibleValue.ZERO);
-        } else {
-            return !info.getCreationtxid().equals(Sha256Hash.ZERO_HASH);
-        }
+    /**
+     * Try to get a property and if successful, put it in the cache
+     * @param id property to fetch
+     */
+    private void getPropertyAsync(CurrencyID id) {
+        log.debug("Fetching {}", id);
+        var disposable = omniGetPropertyMaybe(id).subscribe(cache::cachePut, this::onError);
+    }
+
+    private Maybe<OmniPropertyInfo> omniGetPropertyMaybe(CurrencyID id) {
+        return rxJsonClient.pollOnce(() -> rxJsonClient.omniGetProperty(id));
+    }
+
+    private Maybe<List<SmartPropertyListInfo>> omniListPropertiesMaybe() {
+        return rxJsonClient.pollOnce(rxJsonClient::omniListProperties);
     }
 }
